Split services_client.py into different tests
